create custom hook and clean App.js
